+++ components/AppContext.ts
@@ -187,7 +187,6 @@ import { createExchangeQuote$, ExchangeAction, ExchangeType } from 'features/exc
 import { followedVaults$ } from 'features/follow/api'
 import { createGeneralManageVault$ } from 'features/generalManageVault/generalManageVault'
 import { VaultType } from 'features/generalManageVault/vaultType'
-import { getOasisStats$ } from 'features/homepage/stats'
 import { createIlkDataListWithBalances$ } from 'features/ilks/ilksWithBalances'
 import { createManageMultiplyVault$ } from 'features/multiply/manage/pipes/manageMultiplyVault'
 import { createOpenMultiplyVault$ } from 'features/multiply/open/pipes/openMultiplyVault'
@@ -1343,7 +1342,7 @@ export function setupAppContext() {
     openGuniVault$,
     ilkDataList$,
     connectedContext$,
-    getOasisStats$: memoize(getOasisStats$),
+
     addGasEstimation$,
     instiVault$,
     ilkToToken$,
