+++ components/layouts/LandingPageLayout.tsx
@@ -1,0 +1,27 @@
+import { Footer } from 'components/Footer'
+import { HomepageHero } from 'features/homepage/HomepageHero'
+import { NavigationControllerDynamic } from 'features/navigation/controls/NavigationControllerDynamic'
+import { WithChildren } from 'helpers/types'
+import React from 'react'
+
+import { WithAnnouncementLandingLayout } from './WithAnnouncementLayout'
+
+export function LandingPageLayout({ children }: WithChildren) {
+  return (
+    <>
+      <WithAnnouncementLandingLayout
+        header={
+          <>
+            <NavigationControllerDynamic /> <HomepageHero />
+          </>
+        }
+        footer={<Footer />}
+        showAnnouncement={false}
+        variant="landingContainer"
+        sx={{ position: 'relative', zIndex: 1 }}
+      >
+        {children}
+      </WithAnnouncementLandingLayout>
+    </>
+  )
+}
