+++ components/layouts/backgrounds.tsx
@@ -1,0 +1,9 @@
+import { Background } from 'theme/Background'
+import { BackgroundLight } from 'theme/BackgroundLight'
+
+export const marketingBackgrounds = {
+  default: <Background />,
+  light: <BackgroundLight />,
+  lighter: <BackgroundLight />,
+  none: null,
+}
