+++ features/ajna/common/layout.tsx
@@ -2,7 +2,7 @@ import { isAppContextAvailable } from 'components/AppContextProvider'
 import { WithFeatureToggleRedirect } from 'components/FeatureToggleRedirect'
 import { Footer } from 'components/Footer'
 import { PageSEOTags } from 'components/HeadTags'
-import { WithAnnouncementLayout } from 'components/Layouts'
+import { WithAnnouncementLayout } from 'components/layouts/WithAnnouncementLayout'
 import { AjnaNavigationController } from 'features/navigation/controls/AjnaNavigationController'
 import React, { PropsWithChildren } from 'react'
 import { ajnaExtensionTheme } from 'theme'
