+++ features/homepage/HomepageHero.tsx
@@ -1,9 +1,9 @@
 import BigNumber from 'bignumber.js'
-import { useAppContext } from 'components/AppContextProvider'
 import { NewReferralModal } from 'features/referralOverview/NewReferralModal'
-import { TermsOfService } from 'features/termsOfService/TermsOfService'
+import { checkReferralLocalStorage } from 'features/referralOverview/referralLocal'
+import { TermsOfServiceReferralDynamic } from 'features/termsOfService/TermsOfServiceReferralDynamic'
 import { formatAsShorthandNumbers } from 'helpers/formatters/format'
-import { useObservable } from 'helpers/observableHook'
+import { useAccount } from 'helpers/useAccount'
 import { useFeatureToggle } from 'helpers/useFeatureToggle'
 import { useLocalStorage } from 'helpers/useLocalStorage'
 import { Trans, useTranslation } from 'next-i18next'
@@ -14,6 +14,7 @@ import { Box, Container, Flex, Grid, Text } from 'theme-ui'
 import { Hero } from './common/Hero'
 import { HomepagePromoBlock } from './common/HomepagePromoBlock'
 import { OasisStats } from './OasisStats'
+import { useOasisStats } from './stats'
 
 function StatCell({ label, value }: { label: string; value: string }) {
   return (
@@ -31,7 +32,7 @@ function StatCell({ label, value }: { label: string; value: string }) {
   )
 }
 
-function ManagedVolumeStats({ oasisStatsValue }: { oasisStatsValue?: OasisStats }) {
+function ManagedVolumeStats({ oasisStats }: { oasisStats?: OasisStats }) {
   const { t } = useTranslation()
   return (
     <HomepagePromoBlock.Big
@@ -43,25 +44,25 @@ function ManagedVolumeStats({ oasisStatsValue }: { oasisStatsValue?: OasisStats 
         <StatCell
           label={t('landing.stats.30-day-volume')}
           value={
-            oasisStatsValue
-              ? `$${formatAsShorthandNumbers(new BigNumber(oasisStatsValue.monthlyVolume), 2)}`
+            oasisStats
+              ? `$${formatAsShorthandNumbers(new BigNumber(oasisStats.monthlyVolume), 2)}`
               : '-'
           }
         />
         <StatCell
           label={t('landing.stats.managed-on-oasis')}
           value={
-            oasisStatsValue
-              ? `$${formatAsShorthandNumbers(new BigNumber(oasisStatsValue.managedOnOasis), 2)}`
+            oasisStats
+              ? `$${formatAsShorthandNumbers(new BigNumber(oasisStats.managedOnOasis), 2)}`
               : '-'
           }
         />
         <StatCell
           label={t('landing.stats.collateral-automated')}
           value={
-            oasisStatsValue
+            oasisStats
               ? `$${formatAsShorthandNumbers(
-                  new BigNumber(oasisStatsValue.lockedCollateralActiveTrigger),
+                  new BigNumber(oasisStats.lockedCollateralActiveTrigger),
                   2,
                 )}`
               : '-'
@@ -74,11 +75,9 @@ function ManagedVolumeStats({ oasisStatsValue }: { oasisStatsValue?: OasisStats 
 
 export const HomepageHero = () => {
   const router = useRouter()
-  const { context$, checkReferralLocal$, userReferral$, getOasisStats$ } = useAppContext()
-  const [oasisStatsValue] = useObservable(getOasisStats$())
-  const [context] = useObservable(context$)
-  const [userReferral] = useObservable(userReferral$)
-  const [checkReferralLocal] = useObservable(checkReferralLocal$)
+  const { data: oasisStats } = useOasisStats()
+  const { isConnected } = useAccount()
+  const referralLocal = checkReferralLocalStorage()
 
   useEffect(() => {
     if (!localReferral && referralsEnabled) {
@@ -88,7 +87,7 @@ export const HomepageHero = () => {
         setLandedWithRef(linkReferral)
       }
     }
-  }, [checkReferralLocal, router.isReady])
+  }, [referralLocal, router.isReady])
 
   const referralsEnabled = useFeatureToggle('Referrals')
   const notificationsEnabled = useFeatureToggle('Notifications')
@@ -96,10 +95,8 @@ export const HomepageHero = () => {
   const [localReferral, setLocalReferral] = useLocalStorage('referral', '')
   return (
     <Container>
-      {referralsEnabled && landedWithRef && context?.status === 'connectedReadonly' && (
-        <NewReferralModal />
-      )}
-      {(referralsEnabled || notificationsEnabled) && <TermsOfService userReferral={userReferral} />}
+      {referralsEnabled && landedWithRef && <NewReferralModal />}
+      {(referralsEnabled || notificationsEnabled) && <TermsOfServiceReferralDynamic />}
       <Flex
         sx={{
           height: 'auto',
@@ -107,11 +104,11 @@ export const HomepageHero = () => {
         }}
       >
         <Hero
-          isConnected={context?.status === 'connected'}
+          isConnected={isConnected}
           heading="landing.hero.main.headline"
           subheading={<Trans i18nKey="landing.hero.main.subheader" components={[<br />]} />}
         />
-        <ManagedVolumeStats oasisStatsValue={oasisStatsValue} />
+        <ManagedVolumeStats oasisStats={oasisStats} />
       </Flex>
     </Container>
   )
