+++ features/homepage/stats.ts
@@ -1,12 +1,7 @@
-import { EMPTY, Observable } from 'rxjs'
-import { ajax } from 'rxjs/ajax'
-import { catchError, map } from 'rxjs/operators'
+import { useFetch } from 'usehooks-ts'
 
-import type { OasisStats } from './OasisStats'
+import { OasisStats } from './OasisStats'
 
-export function getOasisStats$(): Observable<OasisStats> {
-  return ajax.get('/api/stats').pipe(
-    map(({ response }) => response),
-    catchError(() => EMPTY),
-  )
+export function useOasisStats() {
+  return useFetch<OasisStats>('/api/stats')
 }
