+++ features/navigation/controls/NavigationControllerDynamic.tsx
@@ -1,0 +1,5 @@
+import dynamic from 'next/dynamic'
+
+export const NavigationControllerDynamic = dynamic(() => import('./NavigationController'), {
+  ssr: false,
+})
