+++ features/referralOverview/referralLocal.ts
@@ -7,3 +7,11 @@ export function checkReferralLocalStorage$(): Observable<string | null> {
   }
   return of(referrer)
 }
+
+export function checkReferralLocalStorage(): string | null {
+  let referrer = localStorage.getItem(`referral`)
+  if (referrer === 'null') {
+    referrer = null
+  }
+  return referrer
+}
