+++ features/termsOfService/TermsOfServiceReferralDynamic.tsx
@@ -1,0 +1,5 @@
+import dynamic from 'next/dynamic'
+
+export const TermsOfServiceReferralDynamic = dynamic(() => import('./TermsOfServiceReferral'), {
+  ssr: false,
+})
